feat: new navigation
